Front LED Apron
The 8 Individual pieces provided by Fysetc did not secure to the 2020 extrusion properly and would tend to fall out. I used the Fysetc provided models to create LED_00-03.stl and LED_04-07.stl. Rather that trying to secure 8 pieces, it is now only 2 pieces, each of which is fastened to the 2020 extrusion.